upgraded to Jedis 1.5.2, added missing methods from Redis bean, changed redis bean to be singleton. Fix for GPREDIS-8 and GPREDIS-9

+++ spring-datastore-redis/src/main/java/org/springframework/datastore/mapping/redis/collection/RedisList.java
@@ -14,14 +14,9 @@
  */
 package org.springframework.datastore.mapping.redis.collection;
 
-import java.util.AbstractList;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Set;
-
 import org.springframework.datastore.mapping.redis.util.RedisTemplate;
+
+import java.util.*;
 
 /**
  * Creates a list that is backed onto a Redis list
@@ -72,7 +67,7 @@ public class RedisList extends AbstractList implements List, RedisCollection {
 
     @Override
     public int size() {
-        return redisTemplate.llen(redisKey);
+        return Long.valueOf(redisTemplate.llen(redisKey)).intValue();
     }
 
     @Override
